Fix: Fix bug for last commit
1. Fix the bug for integer
2. Change reset sequence, change arm position first

+++ src/Robot/MotorUnion/Motor/motor/motor.cpp
@@ -90,7 +90,7 @@ void motor::SetMotor_Angle(const float &angle)
 			Motor_Scale = Min_Extend_Limit;
 		else
 			;
-		Motor_Angle = Motor_Scale / Rev_Resol_Extended * 360;
+		Motor_Angle = (float)Motor_Scale / Rev_Resol_Extended * 360;
 		
 		break;
 	}
